Minimizing current consumption on SIMO startup.
Minimizing current consumption prevents the SIMO from browning out
when in soft start on POR. This can happen when there is bounce on
the power supply. This allows the SIMO to reach the set point and
get out of soft start.

Setting the proper OVR settings on MAX32665 to optimize SRAM timing.

MSDK-Commit: 449756af20ba71b3d5bf3b6a4b7f9f719d928f2c