Fix/adc (#433)
* Fix width of adcfrq field in peripheral clock divisor register

GCR_PCLKDIV.adcfrq is 4 bits wide however it is listed as 3 bits wide in
MAX32655, MAX32680 and MAX78000 header and svd files.

* MSDK-944: Update ADC example to read external analog input

MAX32680 datasheet states that the 10-bit ADC has four external inputs
labelled from AIN12 to AIN15. Their corresponding channels are given as
CH4 to CH7.

Update the example to monitor AIN12,CH4 which is routed to JH5:4 on EvKit.

* Fix adcfrq field in me17 and ai85 register definition files

MSDK-Commit: 27979fe983642c292053b2eb09863a39d1fc1a3a