fix(PeriphDrivers, Other): Fix system clock configuration problems of MAX32650 (#1458)
SystemCoreClock parameter is not Initialized properly for MAX32650 on
Zephyr side. That causes problems during timeout calculations and other
peripheral initializations that depend on system clock value. Fix this issue
by checking the chip revision and setting the SystemCoreClock value
accordingly during system initialization.

When configuring the system clock to use the 32kHz oscillator on the
MAX32650, the code incorrectly set the SYSOSC_SEL field to a ready
flag value instead of the correct selection value.

Signed-off-by: Furkan Akkiz <hasanfurkan.akkiz@analog.com>